Revert "feat: error page"
This reverts commit 1d8207be3d8290f099d693bb9325bd038dbabf3b.
