+++ packages/core/src/app/components/ApplicationFile.tsx
@@ -4,7 +4,6 @@ import { File } from '@shuvi/react-fs';
 function ApplicationFile() {
   const app = `
 import AppComponent from "@shuvi/app/core/app";
-import ErrorComponent from "@shuvi/app/core/error";
 import routes from "@shuvi/app/core/routes";
 import initPlugins from "@shuvi/app/core/plugin";
 import { pluginRecord } from "@shuvi/app/core/plugins";
@@ -23,7 +22,6 @@ export function create(context, options) {
 
   app = new Application({
     AppComponent,
-    ErrorComponent,
     routes,
     context,
     render: options.render
@@ -43,9 +41,8 @@ if (module.hot) {
     if (!app) return;
 
     let AppComponent = require('@shuvi/app/core/app').default;
-    let ErrorComponent = require('@shuvi/app/core/error').default;
     let routes = require('@shuvi/app/core/routes').default;
-    app.rerender({routes,AppComponent,ErrorComponent});
+    app.rerender({routes,AppComponent});
   });
 }
 `;
@@ -56,7 +53,6 @@ import { Application } from "@shuvi/core/lib/app/app-modules/application";
 export function create(context, options) {
   return new Application({
     AppComponent: null,
-    ErrorComponent: null,
     routes: [],
     context,
     render: options.render
