+++ packages/core/src/types/application.ts
@@ -1,21 +1,11 @@
 import { Hookable } from '@shuvi/hooks';
 import { IRoute } from './route';
 
-export type IError = Error | undefined;
-
 export interface IRenderOptions<CompType = any> {
   AppComponent: CompType;
   routes: IRoute[];
   appContext: Record<string, any>;
 }
-
-export type IRenderErrorOptions<CompType = any> = Pick<
-  IRenderOptions,
-  'appContext'
-> & {
-  error: IError;
-  ErrorComponent: CompType;
-};
 
 export type IRerenderConfig = {
   AppComponent?: any;
@@ -28,7 +18,6 @@ export interface IAppRenderFn {
 
 export interface IApplication extends Hookable {
   AppComponent: any;
-  ErrorComponent: any;
   routes: IRoute[];
   run(): Promise<{ [k: string]: any }>;
   rerender(config?: IRerenderConfig): Promise<void>;
