+++ packages/router/src/router.ts
@@ -1,5 +1,3 @@
-/// <reference lib="dom" />
-
 import { History, Location, Listener } from './history';
 
 export interface IRouter {
@@ -15,14 +13,7 @@ export interface IRouter {
   // @internal
   createHref: History['createHref'];
   forward(): void;
-  onChange: (listener: Listener) => () => void;
-}
-
-export function handleRouterError(router: IRouter) {
-  const unsubscribe = router.onChange(({ location }) => {
-    window.location.href = location.pathname;
-    unsubscribe();
-  });
+  onChange: (listener: Listener) => void;
 }
 
 export const createRouter = (history: History): IRouter => {
