+++ packages/runtime-react/src/lib/index.ts
@@ -70,10 +70,6 @@ loadRouteComponent(() => import(/* webpackChunkName: "page-${route.id}" */"${com
     return resolveAppFile('App');
   }
 
-  getErrorModulePath(): string {
-    return resolveAppFile('errorPage');
-  }
-
   get404ModulePath(): string {
     return resolveAppFile('page404');
   }
