+++ packages/runtime-react/src/shuvi-app/router/renderRoutes.tsx
@@ -40,6 +40,10 @@ function renderRoutesChildrens(
       );
     }
 
+    if (route.name === '404') {
+      appContext.statusCode = 404;
+    }
+
     return (
       <Route
         key={route.id || i}
