+++ packages/shuvi/src/api/api.ts
@@ -193,7 +193,7 @@ class Api extends Hookable implements IApi {
     routes.push({
       path: '*',
       component: this.resolveAppFile('core', '404'),
-      name: 'notFound'
+      name: '404'
     });
 
     this._routes = routes;
