+++ packages/shuvi/src/lib/renderToHTML.ts
@@ -1,8 +1,6 @@
 import { Runtime } from '@shuvi/types';
 import { Renderer, isRedirect } from '../renderer';
 import { Api } from '../api';
-import { IError } from '@shuvi/core';
-import { NOT_FOUND_ERROR_MESSAGE } from '@shuvi/shared/lib/constants';
 
 export async function renderToHTML({
   req,
@@ -15,39 +13,10 @@ export async function renderToHTML({
 }): Promise<{ html: string | null; appContext: any }> {
   let html: null | string = null;
   const renderer = new Renderer({ api });
-  const {
-    server: { application, view },
-    clientManifest: manifest
-  } = api.resources;
-  const getAssetPublicUrl = api.getAssetPublicUrl.bind(api);
-
+  const { application } = api.resources.server;
   const app = application.create(
     {
-      req,
-      error: function (error: IError) {
-        if (!this.error.result) {
-          this.error.message = error?.message;
-
-          if (error?.message === NOT_FOUND_ERROR_MESSAGE) {
-            this.statusCode = 404;
-            error = undefined;
-          } else {
-            console.error(error);
-            this.statusCode = 500;
-          }
-
-          this.error.result = view.renderError({
-            error,
-            appContext: this,
-            url: req.url || '/',
-            ErrorComponent: app.ErrorComponent,
-            getAssetPublicUrl,
-            manifest
-          });
-        }
-
-        return this.error.result;
-      }
+      req
     },
     {
       async render({ appContext, AppComponent, routes }) {
