+++ test/fixtures/custom-404/src/404.js
@@ -1,0 +1,3 @@
+export default function Page404() {
+  return <div id="custom-404">404</div>
+}
