--- test/fixtures/custom-error/src/pages/throwErrorInGetInitialProps.js
@@ -1,7 +1,0 @@
-const ThrowError = () => {
-  return 'should not render';
-};
-
-ThrowError.getInitialProps = () => {
-  throw new Error('error');
-};
