--- test/fixtures/custom-error/src/pages/throwErrorInRender.js
@@ -1,3 +1,0 @@
-export default () => {
-  throw new Error('Bad error');
-};
