--- test/fixtures/page-error/shuvi.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  ssr: true
-};
