--- test/fixtures/page-error/src/pages/error.js
@@ -1,3 +1,0 @@
-export default () => {
-  throw new Error('test-error');
-};
