--- test/fixtures/page-error/src/pages/index.js
@@ -1,1 +1,0 @@
-export default () => <div id="index">Index Page</div>;
