+++ test/utils/browser/index.ts
@@ -7,7 +7,6 @@ import puppeteer, {
 } from 'puppeteer-core';
 import qs from 'querystring';
 import ChromeDetector from './chrome';
-import { IRouter } from '@shuvi/router';
 
 export interface Page extends puppeteer.Page {
   [x: string]: any;
@@ -21,7 +20,6 @@ export interface Page extends puppeteer.Page {
   statusCode?: number;
   shuvi: {
     navigate(path: string, query?: Record<string, any>): Promise<any>;
-    invokeRouter: (action: keyof IRouter, arg?: any) => Promise<any>;
   };
 }
 
@@ -139,7 +137,6 @@ export default class Browser {
       );
 
     const getShuvi = () => page.evaluateHandle('window.__SHUVI');
-
     page.shuvi = {
       async navigate(path: string, query: Record<string, any> = {}) {
         const $shuvi = await getShuvi();
@@ -150,15 +147,6 @@ export default class Browser {
           ($shuvi, path: string) => $shuvi.router.push(path),
           $shuvi,
           path
-        );
-      },
-      async invokeRouter(action: keyof IRouter, arg: any) {
-        const $shuvi = await getShuvi();
-        return page.evaluate(
-          ($shuvi, action: keyof IRouter, arg) => $shuvi.router[action](arg),
-          $shuvi,
-          action,
-          arg
         );
       }
     };
